+++ src/py42/_internal/clients/util.py
@@ -6,39 +6,31 @@ import py42.util as util
 def for_each_api_item(
     init_response,
     obj_retriever,
-    foreach_page_size,
-    foreach_user_callback=None,
-    foreach_datakey=None,
-    foreach_return_each_page=False,
+    py42_page_size,
+    py42_callback=None,
+    py42_datakey=None,
+    py42_return_each_page=False,
     **kwargs
 ):
-    total_count = util.get_obj_from_response(init_response, u"totalCount")
-    _foreach_response(
-        init_response, foreach_user_callback, foreach_datakey, foreach_return_each_page, **kwargs
-    )
-    pages = int(math.ceil(float(total_count) / float(foreach_page_size)))
+    def process_page(response):
+        _foreach_response(response, py42_callback, py42_datakey, py42_return_each_page)
+
+    total_count = util.get_obj_from_response(init_response, "totalCount")
+
+    process_page(init_response)
+
+    pages = int(math.ceil(float(total_count) / float(py42_page_size)))
+
     if pages > 1:
         for i in range(1, pages):  # skip the first page since we already handled that above
             page_num = i + 1
-
-            def get_page(response):
-                _foreach_response(
-                    response,
-                    foreach_user_callback,
-                    foreach_datakey,
-                    foreach_return_each_page,
-                    **kwargs
-                )
-
-            obj_retriever(page_num=page_num, page_size=foreach_page_size, then=get_page)
+            obj_retriever(page_num=page_num, page_size=py42_page_size, then=process_page, **kwargs)
 
 
-def _foreach_response(
-    response, foreach_user_callback, foreach_datakey, foreach_return_each_page, **kwargs
-):
-    items = util.get_obj_from_response(response, foreach_datakey)
-    if foreach_return_each_page:
-        foreach_user_callback(items, **kwargs)
+def _foreach_response(response, py42_callback, py42_datakey, py42_return_each_page):
+    items = util.get_obj_from_response(response, py42_datakey)
+    if py42_return_each_page:
+        py42_callback(items)
     else:
         for item in items:
-            foreach_user_callback(item, **kwargs)
+            py42_callback(item)