+++ tests/sdk/conftest.py
@@ -7,7 +7,7 @@ from py42._internal.file_event_filter import FileEventFilter, FilterGroup
 EVENT_FILTER_FIELD_NAME = "filter_field_name"
 OPERATOR_STRING = "IS_IN"
 VALUE_STRING = "value_example"
-VALUE_UNICODE = u"我能吞下玻璃而不伤身体"
+VALUE_UNICODE = u"您已经发现了秘密信息"
 
 
 @pytest.fixture